Moved form-related views into generic framework
[refactoring]

Since the four views that display and evaluate forms had a very
similar structure, I decided to factor out the common parts into a
framework that can be customized by supplying the individual parts in
functions passed as parameters to the "view maker" function.

I did this in the hope of sparing line count and complexity, none of
which has so far payed off, but it is certain that more views will be
added that can use this framework.