models, views: rights, visibility displayed as text
[access control]

User rights are displayed in textual form on heap pages. In order to achieve
this, a class method was added to the `UserRight` model that returns the
display string for arbitrary right IDs (`get_right_display` only works on
instances, whereas no `UserRight` instance belongs to non-asserted rights that
are derived from heap visibility or the user's superuser status).

Also, rights granted to 'everyone else' due to heap visibility are displayed.