Added database integrity check
[fsck]

Database integrity check (a.k.a. 'fsck') is accessible as a view, but
only to superusers. It runs several tests the detect possible
corruptions of the object database. I have written tests for all
errors I could think of, but that very probably does not cover all the
ways a database can go wrong.

Of course, the long-term goal is to have a UI that does not allow
creating an invalid database state.