Facelift and UI for message labels added
[ui feature]

The most prominent feature in this commit is a thorough facelift. CSS
was added, though it now resides in the HTML file, and it should moved
into a separate, statically served stylesheet. The appearance now
includes an industry standard top bar, and parts of the message header
in the conversation view have been reshuffled into a more convenient
layout, now complete with Gravatar icons. In the absence of a Gravatar
icon, several default icons are used w.r.t. the rights of the author
on the heap. These icons are highly provisional, and at the moment,
they are hotlinked to other sites, which is a controversial practice,
so please do not deploy this version without serving these files
yourselves. Thanks.

Besides aesthetic improvements, I also coded two new views and related
UI to implement adding and removing message labels.

Remaining major usability issues:
- navigation, esp. stepping back to the conversation list of a heap
  from a specific conversation (a breadcrumb bar would be nice),
- conversation subjects cannot be altered from the UI,
- there is no way to merge conversations or move messages between
  conversations,
- anonymous users do not see a list of heaps accessible to them, so
  they can only access content via URLs.